refactor(renderer): deepen career profile ownership
